llm/pyproject.toml
[project]
name = "my-img-gen-llm"
version = "0.1.0"
description = "LLM-powered image generation service"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["llm", "image-generation", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
    "diffusers",
    "transformers",
    "torch",
    "pillow",
    "sentencepiece>=0.2.1",
    "sacremoses>=0.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "black>=25.1.0",
    "isort>=6.0.0",
    "mypy>=1.17.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/my-img-gen"
Repository = "https://github.com/yourusername/my-img-gen"
Issues = "https://github.com/yourusername/my-img-gen/issues"

[project.scripts]
my-img-gen-llm = "my_img_gen_llm.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["my_img_gen_llm"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["my_img_gen_llm"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
